+++ x/evm/keeper/block_proposer.go
@@ -6,10 +6,31 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 )
 
-// GetCoinbaseAddress returns the block proposer's validator operator address.
-// TODO ES: (long comment) store this per block height to avoid querying the staking module, as well as reduce number of parameters required for `eth_*` calls. After removed, cleanup the unused functions like `GetProposerAddress`
-func (k Keeper) GetCoinbaseAddress(ctx sdk.Context, proposerAddress sdk.ConsAddress) (common.Address, error) {
-	validator, err := k.stakingKeeper.GetValidatorByConsAddr(ctx, GetProposerAddress(ctx, proposerAddress))
+// GetCoinbaseAddress returns the block proposer's validator operator address from the context proposer address.
+// The proposer address can be overridden with the provided address.
+// If the proposer address is empty, it returns empty address instead of an error.
+func (k Keeper) GetCoinbaseAddress(ctx sdk.Context, overrideProposerAddress sdk.ConsAddress) (common.Address, error) {
+	proposerAddress := sdk.ConsAddress(ctx.BlockHeader().ProposerAddress)
+	if len(overrideProposerAddress) > 0 {
+		proposerAddress = ctx.BlockHeader().ProposerAddress
+	}
+
+	isEmptyProposerAddress := func() bool {
+		if len(proposerAddress) == 0 {
+			return true
+		}
+		for _, b := range proposerAddress {
+			if b != 0 {
+				return false
+			}
+		}
+		return true
+	}()
+	if isEmptyProposerAddress {
+		return common.Address{}, nil
+	}
+
+	validator, err := k.stakingKeeper.GetValidatorByConsAddr(ctx, proposerAddress)
 	if err != nil {
 		return common.Address{}, errorsmod.Wrapf(
 			err,
@@ -26,14 +47,7 @@ func (k Keeper) GetCoinbaseAddress(ctx sdk.Context, proposerAddress sdk.ConsAddr
 			validator.GetOperator(),
 		)
 	}
+
 	coinbase := common.BytesToAddress(valAddr)
 	return coinbase, nil
-}
-
-// GetProposerAddress returns current block proposer's address when provided proposer address is empty.
-func GetProposerAddress(ctx sdk.Context, proposerAddress sdk.ConsAddress) sdk.ConsAddress {
-	if len(proposerAddress) == 0 {
-		proposerAddress = ctx.BlockHeader().ProposerAddress
-	}
-	return proposerAddress
 }