+++ CHANGELOG.md
@@ -51,6 +51,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 - (ante) [#161](https://github.com/EscanBE/evermint/pull/161) Merge `EthBasicValidationDecorator` into `EthValidateBasicDecorator` with enhancement
 - (evm) [#163](https://github.com/EscanBE/evermint/pull/163) Remove support for unprotected txs and cleanup deprecated proto
 - (evm) [#170](https://github.com/EscanBE/evermint/pull/170) Minor improvement
+- (evm) [#171](https://github.com/EscanBE/evermint/pull/171) Refactor code relates to `proposer address` in `x/evm`
 
 ### Bug Fixes